refactor(installs): refactor installs view into hooks
